Moved functions in the snap_info_utility.py
